+++ pkg/analyzer/CHANGELOG.md
@@ -31,6 +31,7 @@
 * Deprecate `InstanceFragment.methods2`, use `methods` instead.
 * Deprecate `InterfaceElement.constructors2`, use `constructors` instead.
 * Deprecate `InterfaceFragment.constructors2`, use `constructors` instead.
+* Deprecate `JoinPatternVariableElement.variables2`, use `variables` instead.
 
 ## 7.4.1
 * Restore `InstanceElement.augmented` getter.